Moved subscription features to network.bookshare.subscription package
Refined strings.xml file
