+++ src/org/geometerplus/android/fbreader/FBReader.java
@@ -36,10 +36,10 @@ import org.geometerplus.android.fbreader.benetech.SpeakActivity;
 import org.geometerplus.android.fbreader.library.KillerCallback;
 import org.geometerplus.android.fbreader.library.SQLiteBooksDatabase;
 import org.geometerplus.android.fbreader.network.bookshare.BookshareDeveloperKey;
-import org.geometerplus.android.fbreader.network.bookshare.SubscriptionAlarmTriggerService;
-import org.geometerplus.android.fbreader.subscription.BooksharePeriodicalDataSource;
-import org.geometerplus.android.fbreader.subscription.PeriodicalEntity;
-import org.geometerplus.android.fbreader.subscription.PeriodicalsSQLiteHelper;
+import org.geometerplus.android.fbreader.network.bookshare.subscription.BooksharePeriodicalDataSource;
+import org.geometerplus.android.fbreader.network.bookshare.subscription.PeriodicalEntity;
+import org.geometerplus.android.fbreader.network.bookshare.subscription.PeriodicalsSQLiteHelper;
+import org.geometerplus.android.fbreader.network.bookshare.subscription.SubscriptionAlarmTriggerService;
 import org.geometerplus.android.fbreader.tips.TipsActivity;
 import org.geometerplus.android.util.UIUtil;
 import org.geometerplus.fbreader.Paths;
