+++ src/org/geometerplus/android/fbreader/network/bookshare/Bookshare_Periodical_Edition_Listing.java
@@ -48,126 +48,135 @@ import android.widget.ListView;
 import android.widget.SimpleAdapter;
 import android.widget.TextView;
 
-public class Bookshare_Periodical_Edition_Listing extends ListActivity{
+public class Bookshare_Periodical_Edition_Listing extends ListActivity {
 
 	protected final static int EDITION_LIST_RESPONSE = 1;
 	protected final static int EDITION_METADATA_RESPONSE = 2;
-	
+
 	private String username;
 	private String password;
 	private String requestURI;
 	private int requestType;
 	private int responseType;
-	
-	protected final static String URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL + Bookshare_Webservice_Login.BOOKSHARE_API_HOST + "/periodical/id/";
-	
+
+	protected final static String URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
+			+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST + "/periodical/id/";
+
 	private final int DATA_FETCHED = 99;
 	private Vector<Bookshare_Periodical_Edition_Bean> vectorResults;
 	private ProgressDialog pd_spinning;
-	
+
 	private final int START_BOOKSHARE_EDITION_DETAILS_ACTIVITY = 13;
 	private final int BOOKSHARE_EDITION_DETAILS_FINISHED = 14;
 	private final int BOOKSHARE_EDITION_LISTING_FINISHED = 11;
 	private final int START_BOOKSHARE_LOGIN_ACTIVITY = 12;
-	
+
 	private final int PREVIOUS_PAGE_BOOK_ID = -1;
 	private final int NEXT_PAGE_BOOK_ID = -2;
 	private int current_result_page = 1;
 	private boolean total_pages_count_known = false;
-	
-	private ArrayList<TreeMap<String,Object>> list = new ArrayList<TreeMap<String, Object>>();
+
+	private ArrayList<TreeMap<String, Object>> list = new ArrayList<TreeMap<String, Object>>();
 	InputStream inputStream;
-	final BookshareWebservice bws = new BookshareWebservice(Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
+	final BookshareWebservice bws = new BookshareWebservice(
+			Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
 	private int total_pages_result;
-	private Boolean isFree=false;
+	private Boolean isFree = false;
 	private String developerKey = BookshareDeveloperKey.DEVELOPER_KEY;
 	private Resources resources;
 	private EditText search_text;
-	
-	private String uri;	//string which contains the uri to fetch periodical details
+
+	private String uri; // string which contains the uri to fetch periodical
+						// details
 	private String bookshare_ID;
 	private String bookshare_edition;
 	private String bookshare_revision;
 	private String bookshare_title;
 	private boolean isOM;
-	
+
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		
+
 		requestWindowFeature(Window.FEATURE_NO_TITLE);
-		
+
 		resources = getApplicationContext().getResources();
 
-		Intent intent  = getIntent();
+		Intent intent = getIntent();
 		username = intent.getStringExtra("username");
 		password = intent.getStringExtra("password");
 
-		if(username == null || password == null){
+		if (username == null || password == null) {
 			isFree = true;
 		}
 
-		// Obtain the application wide SharedPreferences object and store the login information
-		//Get information about user. (i.e. Whether he's an organizational member)
-		SharedPreferences login_preference = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
+		// Obtain the application wide SharedPreferences object and store the
+		// login information
+		// Get information about user. (i.e. Whether he's an organizational
+		// member)
+		SharedPreferences login_preference = PreferenceManager
+				.getDefaultSharedPreferences(getApplicationContext());
 		isOM = login_preference.getBoolean("isOM", false);
-		
-		//Request URI is not needed since we're not using search terms (thushv)
+
+		// Request URI is not needed since we're not using search terms (thushv)
 		requestURI = intent.getStringExtra("ID_SEARCH_URI");
 
-		requestType = intent.getIntExtra(Bookshare_Menu.REQUEST_TYPE, Bookshare_Menu.PERIODICAL_EDITION_REQUEST);
-		responseType= EDITION_LIST_RESPONSE;
+		requestType = intent.getIntExtra(Bookshare_Menu.REQUEST_TYPE,
+				Bookshare_Menu.PERIODICAL_EDITION_REQUEST);
+		responseType = EDITION_LIST_RESPONSE;
 
 		getListing(requestURI);
-		
+
 		getListView().setOnCreateContextMenuListener(this);
 
 	}
 
 	/*
-     * Display voiceable message and then close
-    */
-    private void confirmAndClose(String msg, int timeout) {
-        final ParentCloserDialog dialog = new ParentCloserDialog(this, this);
-        dialog.popup(msg, timeout);
-    }
-    
-    
+	 * Display voiceable message and then close
+	 */
+	private void confirmAndClose(String msg, int timeout) {
+		final ParentCloserDialog dialog = new ParentCloserDialog(this, this);
+		dialog.popup(msg, timeout);
+	}
+
 	private void getListing(final String uri) {
 		vectorResults = new Vector<Bookshare_Periodical_Edition_Bean>();
 
-		//Show progress bar
-		pd_spinning = ProgressDialog.show(this, null, resources.getString(R.string.fetching_periodicals), Boolean.TRUE);
+		// Show progress bar
+		pd_spinning = ProgressDialog.show(this, null,
+				resources.getString(R.string.fetching_periodicals),
+				Boolean.TRUE);
 
-		new Thread(){
-			public void run(){
-				try{
+		new Thread() {
+			public void run() {
+				try {
 					inputStream = bws.getResponseStream(password, uri);
 
-					// Once the response is obtained, send message to the handler
+					// Once the response is obtained, send message to the
+					// handler
 					Message msg = Message.obtain();
 					msg.what = DATA_FETCHED;
 					msg.setTarget(handler);
 					msg.sendToTarget();
-				}catch(IOException ioe){
+				} catch (IOException ioe) {
 					System.out.println(ioe);
-				}
-				catch(URISyntaxException use){
+				} catch (URISyntaxException use) {
 					System.out.println(use);
 				}
 			}
-		}.start();		
+		}.start();
 	}
-	
-	/**When user go to next page, this method clears the list and create the url
-	 *for the next page of results 
-	 * @param selectorId 
+
+	/**
+	 * When user go to next page, this method clears the list and create the url
+	 * for the next page of results
+	 * 
+	 * @param selectorId
 	 */
-	public void pageChangeSelected(int selectorId){
-		if(selectorId == NEXT_PAGE_BOOK_ID){
+	public void pageChangeSelected(int selectorId) {
+		if (selectorId == NEXT_PAGE_BOOK_ID) {
 			current_result_page++;
-		}
-		else if(selectorId == PREVIOUS_PAGE_BOOK_ID){
+		} else if (selectorId == PREVIOUS_PAGE_BOOK_ID) {
 			current_result_page--;
 		}
 		list.clear();
@@ -175,53 +184,58 @@ public class Bookshare_Periodical_Edition_Listing extends ListActivity{
 		StringBuilder strBuilder = new StringBuilder(requestURI);
 		int index;
 
-		if((index = strBuilder.indexOf("?api_key=")) != -1){
+		if ((index = strBuilder.indexOf("?api_key=")) != -1) {
 			strBuilder.delete(index, strBuilder.length());
-			strBuilder.append("/page/"+current_result_page+"?api_key="+developerKey);
+			strBuilder.append("/page/" + current_result_page + "?api_key="
+					+ developerKey);
 		}
 		getListing(strBuilder.toString());
-		
+
 	}
 
-	// Handler for dealing with the stream obtained as a result of search 
-		Handler handler = new Handler(){
+	// Handler for dealing with the stream obtained as a result of search
+	Handler handler = new Handler() {
 
-			@Override
-			public void handleMessage(Message msg){
+		@Override
+		public void handleMessage(Message msg) {
 
-				// Message received that data has been fetched from the bookshare web services 
-				if(msg.what == DATA_FETCHED){
+			// Message received that data has been fetched from the bookshare
+			// web services
+			if (msg.what == DATA_FETCHED) {
 
-					setContentView(R.layout.bookshare_menu_main);
+				setContentView(R.layout.bookshare_menu_main);
 
-					search_text=(EditText)findViewById(R.id.searchText);
-					search_text.setVisibility(View.GONE);
-					
-					// Dismiss the progress dialog
-					pd_spinning.cancel();
+				search_text = (EditText) findViewById(R.id.searchText);
+				search_text.setVisibility(View.GONE);
 
-					String response_HTML = bws.convertStreamToString(inputStream);
+				// Dismiss the progress dialog
+				pd_spinning.cancel();
 
-					// Cleanup the HTML formatted tags
-					String response = response_HTML.replace("&apos;", "\'").replace("&quot;", "\"").replace("&amp;", "and").replace("&#xd;","").replace("&#x97;", "-");
+				String response_HTML = bws.convertStreamToString(inputStream);
 
-					System.out.println(response);
-					// Parse the response of search result
-					parseResponse(response);
+				// Cleanup the HTML formatted tags
+				String response = response_HTML.replace("&apos;", "\'")
+						.replace("&quot;", "\"").replace("&amp;", "and")
+						.replace("&#xd;", "").replace("&#x97;", "-");
 
-					if(responseType==EDITION_METADATA_RESPONSE){
-						//do nothing
-					}
-					if(responseType == EDITION_LIST_RESPONSE){
-					//process list
+				System.out.println(response);
+				// Parse the response of search result
+				parseResponse(response);
+
+				if (responseType == EDITION_METADATA_RESPONSE) {
+					// do nothing
+				}
+				if (responseType == EDITION_LIST_RESPONSE) {
+					// process list
 					list.clear();
 
-					// For each bean object stored in the vector, create a row in the list
-					for(Bookshare_Periodical_Edition_Bean bean : vectorResults){
+					// For each bean object stored in the vector, create a row
+					// in the list
+					for (Bookshare_Periodical_Edition_Bean bean : vectorResults) {
 
 						TreeMap<String, Object> row_item = new TreeMap<String, Object>();
 
-						//puts item title
+						// puts item title
 						row_item.put("title", bean.getTitle());
 						row_item.put("edition", bean.getEdition());
 						row_item.put("revision", bean.getRevision());
@@ -231,414 +245,450 @@ public class Bookshare_Periodical_Edition_Listing extends ListActivity{
 						list.add(row_item);
 					}
 
-					if(current_result_page > 1 ){
-						createPageChanger("Previous Page", PREVIOUS_PAGE_BOOK_ID, R.drawable.arrow_left_blue);
+					if (current_result_page > 1) {
+						createPageChanger("Previous Page",
+								PREVIOUS_PAGE_BOOK_ID,
+								R.drawable.arrow_left_blue);
 					}
 
-					if(current_result_page < total_pages_result ){
-						createPageChanger("Next Page", NEXT_PAGE_BOOK_ID, R.drawable.arrow_right_blue);
-					}
+					if (current_result_page < total_pages_result) {
+						createPageChanger("Next Page", NEXT_PAGE_BOOK_ID,
+								R.drawable.arrow_right_blue);
 					}
 				}
+			}
 
+			// Instantiate the custom SimpleAdapter for populating the ListView
+			// The bookId view in the layout file is used to store the id , but
+			// is not shown on screen
+			MySimpleAdapter simpleadapter = new MySimpleAdapter(
+					getApplicationContext(), list,
+					R.layout.bookshare_menu_item, new String[] { "title",
+							"edition", "download_icon", "book_id" }, new int[] {
+							R.id.text1, R.id.text2,
+							R.id.bookshare_download_icon, R.id.bookId });
 
-				// Instantiate the custom SimpleAdapter for populating the ListView
-				// The bookId view in the layout file is used to store the id , but is not shown on screen
-				MySimpleAdapter simpleadapter = new MySimpleAdapter(
-						getApplicationContext(),list,
-						R.layout.bookshare_menu_item,
-						new String[]{"title","edition","download_icon","book_id"},
-						new int[]{R.id.text1,R.id.text2, R.id.bookshare_download_icon,R.id.bookId});
+			// Set the adapter for this view
+			setListAdapter(simpleadapter);
 
-				
-				
-				//Set the adapter for this view
-				setListAdapter(simpleadapter);
-				
-				ListView lv = getListView();
-				lv.setLongClickable(true);
-				registerForContextMenu(lv);
-				final AlertDialog alert=createSignupDialogBox();
-				
-				View decorView = getWindow().getDecorView();
-				if (null != decorView) {
-					String resultsMessage;
-					if (vectorResults.isEmpty()) {
-						resultsMessage = resources.getString(R.string.search_complete_no_books);
-						setResult(InternalReturnCodes.NO_BOOKS_FOUND);
-						confirmAndClose("no books found", 3000);
-					} else {
-						resultsMessage = resources.getString(R.string.search_complete_with_periodicals);
+			ListView lv = getListView();
+			lv.setLongClickable(true);
+			registerForContextMenu(lv);
+			final AlertDialog alert = createSignupDialogBox();
+
+			View decorView = getWindow().getDecorView();
+			if (null != decorView) {
+				String resultsMessage;
+				if (vectorResults.isEmpty()) {
+					resultsMessage = resources
+							.getString(R.string.search_complete_no_periodicals);
+					setResult(InternalReturnCodes.NO_PERIODICALS_FOUND);
+					confirmAndClose("no periodicals found", 3000);
+				} else {
+					resultsMessage = resources
+							.getString(R.string.search_complete_with_periodicals);
+				}
+				decorView.setContentDescription(resultsMessage);
+			}
+
+			// When list item is clicked item should do a lookup and return
+			// editions/revisions
+			// they have on the particular magazine (thush)
+			lv.setOnItemClickListener(new OnItemClickListener() {
+
+				public void onItemClick(AdapterView<?> parent, View view,
+						int position, long id) {
+
+					// Obtain the layout for selected row
+					LinearLayout row_view = (LinearLayout) view;
+
+					// Obtain the book ID
+					TextView bookId = (TextView) row_view
+							.findViewById(R.id.bookId);
+					TextView periodicalEdition = (TextView) row_view
+							.findViewById(R.id.text2);
+
+					if (null != bookId.getText().toString()) {
+						int numericId = Integer.valueOf(bookId.getText()
+								.toString());
+						if (numericId < 0) {
+							pageChangeSelected(numericId);
+							return;
+						}
 					}
-					decorView.setContentDescription(resultsMessage);
+
+					// Find the corresponding bean object for this row
+					for (Bookshare_Periodical_Edition_Bean bean : vectorResults) {
+
+						// TODO: Fix bug
+						// Since book ID is unique, that can serve as comparison
+						// parameter
+						// Retrieve the book ID from the entry that is clicked
+
+						StringBuilder editionBuilder = new StringBuilder(bean
+								.getEdition());
+						editionBuilder.insert(2, "/");
+						editionBuilder.insert(5, "/");
+						editionBuilder.insert(0, "Edition: ");
+						editionBuilder.append("\t Revision: "
+								+ bean.getRevision());
+
+						if (editionBuilder.toString().equalsIgnoreCase(
+								periodicalEdition.getText().toString().trim())) {
+							bookshare_ID = bean.getId();
+							bookshare_edition = bean.getEdition();
+							bookshare_revision = bean.getRevision();
+							bookshare_title = bean.getTitle();
+
+							if (isFree)
+								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH
+										+ bookshare_ID + "/edition/"
+										+ bookshare_edition + "/revision/"
+										+ bookshare_revision + "?api_key="
+										+ developerKey;
+							else
+								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH
+										+ bookshare_ID + "/edition/"
+										+ bookshare_edition + "/revision/"
+										+ bookshare_revision + "/for/"
+										+ username + "?api_key=" + developerKey;
+
+							if (isOM) {
+								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH
+										+ bookshare_ID + "/edition/"
+										+ bookshare_edition + "/revision/"
+										+ bookshare_revision + "?api_key="
+										+ developerKey;
+							}
+							/*--------------------------------------------------------------------
+							if((isFree && bean.getAvailableToDownload().equals("1") &&
+									bean.getFreelyAvailable().equals("1")) ||
+									(!isFree && bean.getAvailableToDownload().equals("1"))){
+								intent.putExtra("isDownloadable", true);
+							}
+							else{
+								intent.putExtra("isDownloadable", false);
+							}
+							--------------------------------------------------------------------*/
+
+							// intent.putExtra("ID_SEARCH_URI", uri);
+							if (!isFree) {
+								Intent intent = new Intent(
+										getApplicationContext(),
+										Bookshare_Periodical_Edition_Details.class);
+								intent.putExtra("username", username);
+								intent.putExtra("password", password);
+								intent.putExtra("ID_SEARCH_URI", uri);
+								intent.putExtra("PERIODICAL_TITLE",
+										bookshare_title);
+								intent.putExtra("PERIODICAL_ID", bookshare_ID);
+								startActivityForResult(intent,
+										START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
+
+								// It might be not to give user suprises by
+								// having a different view for periodicals
+								// So thought of making it the same
+								// getListView().showContextMenuForChild(view);
+							} else {
+								alert.show();
+
+							}
+
+							// startActivityForResult(intent,
+							// START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
+							break;
+						}
+					}
+
+					// Show a context menu for a single click on list item
+
 				}
 
-				//When list item is clicked item should do a lookup and return editions/revisions
-				//they have on the particular magazine (thush)
-				lv.setOnItemClickListener(new OnItemClickListener(){
+			});
 
-					public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+			lv.setOnItemLongClickListener(new OnItemLongClickListener() {
 
-						// Obtain the layout for selected row
-						LinearLayout row_view  = (LinearLayout)view;
+				@Override
+				public boolean onItemLongClick(AdapterView<?> arg0, View arg1,
+						int arg2, long arg3) {
+					return false;
+				}
+			});
+		}
 
-						//Obtain the book ID
-						TextView bookId = (TextView)row_view.findViewById(R.id.bookId);
-						TextView periodicalEdition = (TextView)row_view.findViewById(R.id.text2);
-						
-						if (null != bookId.getText().toString() ) {
-							int numericId =  Integer.valueOf(bookId.getText().toString());
-							if (numericId < 0) {
-								pageChangeSelected(numericId);
-								return;
-							}
-						}
+		private void createPageChanger(String title, int id, int iconId) {
+			TreeMap<String, Object> row_item = new TreeMap<String, Object>();
+			row_item.put("title", title);
+			row_item.put("edition", "");
 
-						// Find the corresponding bean object for this row
-						for(Bookshare_Periodical_Edition_Bean bean : vectorResults){
+			row_item.put("book_id", String.valueOf(id));
+			row_item.put("download_icon", iconId);
+			list.add(row_item);
+		}
+	};
 
-							//TODO: Fix bug
-							// Since book ID is unique, that can serve as comparison parameter
-							// Retrieve the book ID from the entry that is clicked
-							
-							StringBuilder editionBuilder = new StringBuilder(bean.getEdition());
-							editionBuilder.insert(2, "/");
-							editionBuilder.insert(5, "/");
-							editionBuilder.insert(0, "Edition: ");
-							editionBuilder.append("\t Revision: "+bean.getRevision());
-							
-							if(editionBuilder.toString().equalsIgnoreCase(periodicalEdition.getText().toString().trim())){
-								bookshare_ID = bean.getId();
-								bookshare_edition=bean.getEdition();
-								bookshare_revision=bean.getRevision();
-								bookshare_title=bean.getTitle();
-								
-								
-								if(isFree)
-									uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH + bookshare_ID + "/edition/" + bookshare_edition + "/revision/" + bookshare_revision + "?api_key="+developerKey;
-								else
-									uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH + bookshare_ID + "/edition/" + bookshare_edition + "/revision/" + bookshare_revision + "/for/"+username+"?api_key="+developerKey;
-
-								if(isOM){
-									uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH + bookshare_ID + "/edition/" + bookshare_edition + "/revision/" + bookshare_revision +"?api_key="+developerKey;
-								}
-								/*--------------------------------------------------------------------
-								if((isFree && bean.getAvailableToDownload().equals("1") &&
-										bean.getFreelyAvailable().equals("1")) ||
-										(!isFree && bean.getAvailableToDownload().equals("1"))){
-									intent.putExtra("isDownloadable", true);
-								}
-								else{
-									intent.putExtra("isDownloadable", false);
-								}
-								--------------------------------------------------------------------*/
-								
-						
-								//intent.putExtra("ID_SEARCH_URI", uri);
-								if(!isFree){
-									Intent intent = new Intent(getApplicationContext(),Bookshare_Periodical_Edition_Details.class);
-									intent.putExtra("username", username);
-									intent.putExtra("password", password);
-									intent.putExtra("ID_SEARCH_URI", uri);
-									intent.putExtra("PERIODICAL_TITLE", bookshare_title);
-									intent.putExtra("PERIODICAL_ID", bookshare_ID);
-									startActivityForResult(intent, START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
-									
-									//It might be not to give user suprises by having a different view for periodicals
-									//So thought of making it the same
-									//getListView().showContextMenuForChild(view);
-								}else{
-									alert.show();
-
-								}
-								
-								//startActivityForResult(intent, START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
-								break;
-							}
-						}
-						
-						//Show a context menu for a single click on list item
-						
-					}
-					
-					
-				});
-				
-				lv.setOnItemLongClickListener(new OnItemLongClickListener() {
-
+	// This box will say that user is have not yet signed in. In order to
+	// download
+	// periodicals user must log in.
+	private AlertDialog createSignupDialogBox() {
+		AlertDialog.Builder builder = new AlertDialog.Builder(this);
+		builder.setCancelable(true);
+		// builder.setIcon(android.R.drawable.dialog_question);
+		builder.setTitle("Login Required");
+		builder.setMessage(R.string.login_to_continue);
+		builder.setInverseBackgroundForced(true);
+		builder.setPositiveButton("Login",
+				new DialogInterface.OnClickListener() {
+					// Take the user to the login page with 'continue without
+					// logging in' disabled
 					@Override
-					public boolean onItemLongClick(AdapterView<?> arg0,
-							View arg1, int arg2, long arg3) {
-						return false;
+					public void onClick(DialogInterface dialog, int which) {
+						Intent intent = new Intent(getApplicationContext(),
+								Bookshare_Webservice_Login.class);
+						intent.putExtra("disable_no_login", true);
+						intent.putExtra("periodical_edition_id", bookshare_ID);
+						intent.putExtra("periodical_edition", bookshare_edition);
+						intent.putExtra("periodical_revision",
+								bookshare_revision);
+						intent.putExtra("PERIODICAL_TITLE", bookshare_title);
+						// we need to send this to login page. So when it sees
+						// the request type it can directly send
+						// user to the edition details page
+						intent.putExtra(Bookshare_Menu.REQUEST_TYPE,
+								EDITION_METADATA_RESPONSE);
+						dialog.dismiss();
+
+						startActivityForResult(intent,
+								START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
 					}
 				});
-			}
+		// Just close the dialog box and activity
+		builder.setNegativeButton("Cancel",
+				new DialogInterface.OnClickListener() {
+					@Override
+					public void onClick(DialogInterface dialog, int which) {
+						dialog.dismiss();
+						// finish();
+					}
+				});
+		AlertDialog alert = builder.create();
+		return alert;
 
+	}
 
-			
-
-			
-			private void createPageChanger(String title, int id, int iconId) {
-				TreeMap<String, Object> row_item = new TreeMap<String, Object>();
-				row_item.put("title", title);
-				row_item.put("edition", "");
-			
-				row_item.put("book_id", String.valueOf(id));
-				row_item.put("download_icon", iconId);
-				list.add(row_item);
-			}
-		};
-		
-		//This box will say that user is have not yet signed in. In order to download
-		//periodicals user must log in.
-		private AlertDialog createSignupDialogBox(){
-			AlertDialog.Builder builder = new AlertDialog.Builder(this);
-			builder.setCancelable(true);
-			//builder.setIcon(android.R.drawable.dialog_question);
-			builder.setTitle("Login Required");
-			builder.setMessage(R.string.login_to_continue);
-			builder.setInverseBackgroundForced(true);
-			builder.setPositiveButton("Login", new DialogInterface.OnClickListener() {
-				//Take the user to the login page with 'continue without logging in' disabled
-			  @Override
-			  public void onClick(DialogInterface dialog, int which) {
-				  Intent intent = new Intent(getApplicationContext(),Bookshare_Webservice_Login.class);
-					intent.putExtra("disable_no_login", true);
-					intent.putExtra("periodical_edition_id", bookshare_ID);
-					intent.putExtra("periodical_edition", bookshare_edition);
-					intent.putExtra("periodical_revision", bookshare_revision);
-					intent.putExtra("PERIODICAL_TITLE", bookshare_title);
-					//we need to send this to login page. So when it sees the request type it can directly send
-					//user to the edition details page
-					intent.putExtra(Bookshare_Menu.REQUEST_TYPE, EDITION_METADATA_RESPONSE);
-			    dialog.dismiss();
-			    
-			    startActivityForResult(intent, START_BOOKSHARE_EDITION_DETAILS_ACTIVITY);
-			  }
-			});
-			//Just close the dialog box and activity
-			builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-			  @Override
-			  public void onClick(DialogInterface dialog, int which) {
-			    dialog.dismiss();
-			    //finish();
-			  }
-			});
-			AlertDialog alert = builder.create();
-			return alert;
-			
-		}
-		
-		
-		@Override
-		protected void onActivityResult(int requestCode, int resultCode, Intent data){
-			if(requestCode == START_BOOKSHARE_EDITION_DETAILS_ACTIVITY){
-				if(resultCode == BOOKSHARE_EDITION_DETAILS_FINISHED){
-					setResult(BOOKSHARE_EDITION_LISTING_FINISHED);
-					finish();
-				} else if (resultCode == InternalReturnCodes.NO_BOOK_FOUND) {
-	                setResult(resultCode);
-	                finish();
-	            }
+	@Override
+	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+		if (requestCode == START_BOOKSHARE_EDITION_DETAILS_ACTIVITY) {
+			if (resultCode == BOOKSHARE_EDITION_DETAILS_FINISHED) {
+				setResult(BOOKSHARE_EDITION_LISTING_FINISHED);
+				finish();
+			} else if (resultCode == InternalReturnCodes.NO_PERIODICALS_FOUND) {
+				setResult(resultCode);
+				finish();
 			}
 		}
+	}
 
-		// Used for keeping the the screen from rotating
-		@Override
-		public void onConfigurationChanged(Configuration newConfig){
-			super.onConfigurationChanged(newConfig);
+	// Used for keeping the the screen from rotating
+	@Override
+	public void onConfigurationChanged(Configuration newConfig) {
+		super.onConfigurationChanged(newConfig);
+	}
+
+	/**
+	 * Uses a SAX parser to parse the response
+	 * 
+	 * @param response
+	 *            String representing the response
+	 */
+	private void parseResponse(String response) {
+
+		InputSource is = new InputSource(new StringReader(response));
+
+		try {
+			/* Get a SAXParser from the SAXPArserFactory. */
+			SAXParserFactory spf = SAXParserFactory.newInstance();
+			SAXParser sp;
+			sp = spf.newSAXParser();
+
+			/* Get the XMLReader of the SAXParser we created. */
+			XMLReader parser = sp.getXMLReader();
+			parser.setContentHandler(new SAXHandler());
+			parser.parse(is);
+		} catch (SAXException e) {
+			System.out.println(e);
+		} catch (ParserConfigurationException e) {
+			System.out.println(e);
+		} catch (IOException ioe) {
+			System.out.println(ioe);
+		}
+	}
+
+	// Class containing the logic for parsing the response of search results
+	private class SAXHandler extends DefaultHandler {
+
+		boolean result = false;
+		boolean id = false;
+		boolean title = false;
+		boolean edition = false;
+		boolean revision = false;
+		boolean num_pages = false;
+
+		boolean editionElementVisited = false;
+		boolean revisionElementVisited = false;
+
+		String editionStr;
+
+		Bookshare_Periodical_Edition_Bean result_bean;
+
+		public void startElement(String namespaceURI, String localName,
+				String qName, Attributes atts) {
+
+			if (!total_pages_count_known) {
+				if (qName.equalsIgnoreCase("num-pages")) {
+					num_pages = true;
+					total_pages_count_known = false;
+				}
+			}
+
+			if (qName.equalsIgnoreCase("result")) {
+				result = true;
+				result_bean = new Bookshare_Periodical_Edition_Bean();
+				editionElementVisited = false;
+				revisionElementVisited = false;
+
+			}
+			if (qName.equalsIgnoreCase("id")) {
+				id = true;
+			}
+			if (qName.equalsIgnoreCase("title")) {
+				title = true;
+			}
+			if (qName.equalsIgnoreCase("edition")) {
+				edition = true;
+				if (!editionElementVisited) {
+					editionElementVisited = true;
+				}
+			}
+			if (qName.equalsIgnoreCase("revision")) {
+				revision = true;
+				if (!revisionElementVisited) {
+					revisionElementVisited = true;
+				}
+			}
 		}
 
-		
-		/**
-		 * Uses a SAX parser to parse the response
-		 * @param response String representing the response
+		public void endElement(String uri, String localName, String qName) {
+
+			if (num_pages) {
+				if (qName.equalsIgnoreCase("num-pages")) {
+					num_pages = false;
+				}
+			}
+			if (qName.equalsIgnoreCase("result")) {
+				result = false;
+				vectorResults.add(result_bean);
+				result_bean = null;
+			}
+			if (qName.equalsIgnoreCase("id")) {
+				id = false;
+			}
+			if (qName.equalsIgnoreCase("title")) {
+				title = false;
+			}
+			if (qName.equalsIgnoreCase("edition")) {
+				edition = false;
+			}
+			if (qName.equalsIgnoreCase("revision")) {
+				revision = false;
+			}
+
+		}
+
+		public void characters(char[] c, int start, int length) {
+
+			if (num_pages) {
+				total_pages_result = Integer.parseInt(new String(c, start,
+						length));
+			}
+			if (result) {
+				if (id) {
+					result_bean.setId(new String(c, start, length));
+				}
+				if (title) {
+					result_bean.setTitle(new String(c, start, length));
+				}
+				if (edition) {
+					result_bean.setEdition(new String(c, start, length));
+				}
+				if (revision) {
+					result_bean.setRevision(new String(c, start, length));
+				}
+
+			}
+		}
+	}
+
+	// Here we change the textview2 view to have both edition and revision
+	// strings concatenated
+	// A custom SimpleAdapter class for providing data to the ListView
+	private class MySimpleAdapter extends SimpleAdapter {
+		public MySimpleAdapter(Context context,
+				List<? extends Map<String, ?>> data, int resource,
+				String[] from, int[] to) {
+			super(context, data, resource, from, to);
+		}
+
+		/*
+		 * Retrieves view for the item in the adapter, at the specified position
+		 * and populates it with data.
 		 */
-		private void parseResponse(String response){
+		@Override
+		public View getView(int position, View convertView, ViewGroup parent) {
+			if (convertView == null) {
+				convertView = getLayoutInflater().inflate(
+						R.layout.bookshare_menu_item, null);
 
-			InputSource is = new InputSource(new StringReader(response));
+			}
 
-			try{
-				/* Get a SAXParser from the SAXPArserFactory. */
-				SAXParserFactory spf = SAXParserFactory.newInstance();
-				SAXParser sp;
-				sp = spf.newSAXParser();
+			TreeMap<String, Object> data = (TreeMap<String, Object>) getItem(position);
 
-				/* Get the XMLReader of the SAXParser we created. */
-				XMLReader parser = sp.getXMLReader();
-				parser.setContentHandler(new SAXHandler());
-				parser.parse(is);
+			((TextView) convertView.findViewById(R.id.text1))
+					.setText((String) data.get("title"));
+
+			// Edition and Revision data must be combined and shown in one
+			// textview
+			// Therefore we need to process 2 strings (edition and revision) and
+			// combine 2 strings correctly
+			StringBuilder editionRevisionBuilder = new StringBuilder("");
+			if ((data.get("edition") != null)
+					&& ((String) data.get("edition")).length() > 0) {
+				editionRevisionBuilder = new StringBuilder("Edition: ");
+				String edition = (String) data.get("edition");
+
+				StringBuilder editionBuilder = new StringBuilder(edition);
+				if (edition != null) {
+					editionBuilder.insert(2, "/");
+					editionBuilder.insert(5, "/");
+				}
+
+				editionRevisionBuilder.append(editionBuilder.toString());
+				editionRevisionBuilder.append("\t Revision: ");
+				editionRevisionBuilder.append(((String) data.get("revision")));
+
 			}
-			catch(SAXException e){
-				System.out.println(e);
-			}
-			catch (ParserConfigurationException e) {
-				System.out.println(e);
-			}
-			catch(IOException ioe){
-				System.out.println(ioe);
-			}
+
+			// set the TextViews with appropriate texts
+			String editionR = editionRevisionBuilder.toString();
+			// would have preferred to set this as setContentDescription, but
+			// that didn't voice
+			((TextView) convertView.findViewById(R.id.text2))
+					.setText(editionRevisionBuilder.toString());
+
+			((TextView) convertView.findViewById(R.id.bookId))
+					.setText((String) data.get("book_id"));
+
+			return convertView;
 		}
-
-		// Class containing the logic for parsing the response of search results
-		private class SAXHandler extends DefaultHandler{
-
-			boolean result = false;
-			boolean id = false;
-			boolean title = false;
-			boolean edition = false;
-			boolean revision= false;
-			boolean num_pages = false;
-
-			boolean editionElementVisited = false;
-			boolean revisionElementVisited = false;
-			
-			String editionStr;
-
-			Bookshare_Periodical_Edition_Bean result_bean;
-
-			public void startElement(String namespaceURI, String localName, String qName, Attributes atts){
-				
-				if(!total_pages_count_known)
-				{
-					if(qName.equalsIgnoreCase("num-pages")){
-						num_pages = true;
-						total_pages_count_known = false;
-					}
-				}
-
-				if(qName.equalsIgnoreCase("result")){
-					result = true;
-					result_bean = new Bookshare_Periodical_Edition_Bean();
-					editionElementVisited = false;
-					revisionElementVisited = false;
-			
-				}
-				if(qName.equalsIgnoreCase("id")){
-					id = true;
-				}
-				if(qName.equalsIgnoreCase("title")){
-					title = true;
-				}
-				if(qName.equalsIgnoreCase("edition")){
-					edition = true;
-					if(!editionElementVisited){
-						editionElementVisited = true;
-					}
-				}
-				if(qName.equalsIgnoreCase("revision")){
-					revision = true;
-					if(!revisionElementVisited){
-						revisionElementVisited = true;
-					}
-				}
-			}
-
-			public void endElement(String uri, String localName, String qName){
-
-				if(num_pages){
-					if(qName.equalsIgnoreCase("num-pages")){
-						num_pages = false;
-					}
-				}
-				if(qName.equalsIgnoreCase("result")){
-					result = false;
-					vectorResults.add(result_bean);
-					result_bean = null;
-				}
-				if(qName.equalsIgnoreCase("id")){
-					id = false;
-				}
-				if(qName.equalsIgnoreCase("title")){
-					title = false;
-				}
-				if(qName.equalsIgnoreCase("edition")){
-					edition = false;
-				}
-				if(qName.equalsIgnoreCase("revision")){
-					revision = false;
-				}
-				
-			}
-
-			public void characters(char[] c, int start, int length){
-				
-				if(num_pages){
-					total_pages_result = Integer.parseInt(new String(c,start,length));
-				}
-				if(result){
-					if(id){
-						result_bean.setId(new String(c,start,length));
-					}
-					if(title){
-						result_bean.setTitle(new String(c,start,length));
-					}
-					if(edition){
-						result_bean.setEdition(new String(c,start,length));
-					}
-					if(revision){
-						result_bean.setRevision(new String(c,start,length));
-					}
-					
-				}
-			}
-		}
-
-
-		//Here we change the textview2 view to have both edition and revision strings concatenated
-		// A custom SimpleAdapter class for providing data to the ListView
-		private class MySimpleAdapter extends SimpleAdapter{
-			public MySimpleAdapter(Context context, List<? extends Map<String, ?>> data,
-					int resource, String[] from, int[] to) {
-				super(context, data, resource, from, to);
-			}
-
-			/*
-			 * Retrieves view for the item in the adapter, at the
-			 * specified position and populates it with data.
-			 */
-			@Override
-			public View getView(int position, View convertView, ViewGroup parent) {
-				if (convertView == null) {
-					convertView = getLayoutInflater().inflate(R.layout.bookshare_menu_item, null);
-
-				}
-
-				TreeMap<String, Object> data = (TreeMap<String, Object>) getItem(position);
-
-				((TextView) convertView.findViewById(R.id.text1))
-				.setText((String) data.get("title"));
-
-				//Edition and Revision data must be combined and shown in one textview
-				//Therefore we need to process 2 strings (edition and revision) and
-				//combine 2 strings correctly
-	            StringBuilder editionRevisionBuilder = new StringBuilder("");
-	            if ( (data.get("edition") != null) &&   ((String)data.get("edition")).length() > 0) {
-	                editionRevisionBuilder = new StringBuilder("Edition: ");
-	                String edition=(String)data.get("edition");
-	                
-	                StringBuilder editionBuilder=new StringBuilder(edition);
-	                if(edition!=null){	                	
-	                	editionBuilder.insert(2, "/");
-	                	editionBuilder.insert(5, "/");
-	                }
-	                
-	                editionRevisionBuilder.append(editionBuilder.toString());
-	                editionRevisionBuilder.append("\t Revision: ");
-	                editionRevisionBuilder.append(((String)data.get("revision")));
-	               
-	            }
-
-	            //set the TextViews with appropriate texts
-	            String editionR=editionRevisionBuilder.toString();
-	            // would have preferred to set this as setContentDescription, but that didn't voice
-				((TextView) convertView.findViewById(R.id.text2))
-				.setText(editionRevisionBuilder.toString());
-
-				((TextView) convertView.findViewById(R.id.bookId))
-				.setText((String) data.get("book_id"));
-				
-				return convertView;
-			}
-		}
+	}
 }
