+++ src/org/geometerplus/android/fbreader/network/bookshare/Bookshare_Periodical_Listing.java
@@ -5,9 +5,7 @@ import java.io.InputStream;
 import java.io.StringReader;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.TreeMap;
 import java.util.Vector;
 
@@ -40,6 +38,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.AdapterView;
+import android.widget.AdapterView.OnItemClickListener;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Filter;
@@ -47,20 +46,17 @@ import android.widget.Filterable;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.ListView;
-import android.widget.SimpleAdapter;
 import android.widget.TextView;
-import android.widget.AdapterView.OnItemClickListener;
 
-
-public class Bookshare_Periodical_Listing extends ListActivity{
+public class Bookshare_Periodical_Listing extends ListActivity {
 
 	private final static int LIST_RESPONSE = 1;
 	private final static int METADATA_RESPONSE = 2;
 
-	private String URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL + Bookshare_Webservice_Login.BOOKSHARE_API_HOST + "/periodical/id/";
+	private String URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
+			+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST + "/periodical/id/";
 
-	//Request Response codes
-
+	// Request Response codes
 
 	private String username;
 	private String password;
@@ -73,75 +69,78 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 	private final int START_BOOKSHARE_PERIODICAL_EDITION_ACTIVITY = 10;
 	private final int BOOKSHARE_PERIODICAL_EDITION_FINISHED = 11;
 	private final int BOOKSHARE_PERIODICAL_LISTING_FINISHED = 12;
-	
 
 	private final int PREVIOUS_PAGE_BOOK_ID = -1;
 	private final int NEXT_PAGE_BOOK_ID = -2;
 	private int current_result_page = 1;
 	private boolean total_pages_count_known = false;
-	private ArrayList<TreeMap<String,Object>> list = new ArrayList<TreeMap<String, Object>>();
+	private ArrayList<TreeMap<String, Object>> list = new ArrayList<TreeMap<String, Object>>();
 	InputStream inputStream;
-	final BookshareWebservice bws = new BookshareWebservice(Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
+	final BookshareWebservice bws = new BookshareWebservice(
+			Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
 	private int total_pages_result;
-	private Boolean isFree=false;
+	private Boolean isFree = false;
 	private String developerKey = BookshareDeveloperKey.DEVELOPER_KEY;
 	private Resources resources;
 
 	private EditText searchET;
 
-
-
 	@Override
-	protected void onCreate(Bundle savedInstanceState){
+	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		requestWindowFeature(Window.FEATURE_NO_TITLE);
-		
-		//ListView lv=getListView();
-		//lv.setTextFilterEnabled(true);
+
+		// ListView lv=getListView();
+		// lv.setTextFilterEnabled(true);
 
 		resources = getApplicationContext().getResources();
 
-		Intent intent  = getIntent();
+		Intent intent = getIntent();
 		username = intent.getStringExtra("username");
 		password = intent.getStringExtra("password");
 
-		if(username == null || password == null){
+		if (username == null || password == null) {
 			isFree = true;
 		}
 
-		//Request URI is not needed since we're not using search terms (thushv)
+		// Request URI is not needed since we're not using search terms (thushv)
 		requestURI = intent.getStringExtra(Bookshare_Menu.REQUEST_URI);
 
-		requestType = intent.getIntExtra(Bookshare_Menu.REQUEST_TYPE, Bookshare_Menu.ALL_PERIODICAL_REQUEST);
-		responseType= LIST_RESPONSE;
+		requestType = intent.getIntExtra(Bookshare_Menu.REQUEST_TYPE,
+				Bookshare_Menu.ALL_PERIODICAL_REQUEST);
+		responseType = LIST_RESPONSE;
 
 		getListing(requestURI);
 	}
 
-
-	/**This method fetch data returned from a certain URI
-	 * @param requestURI - The URI to bring data from
+	/**
+	 * This method fetch data returned from a certain URI
+	 * 
+	 * @param requestURI
+	 *            - The URI to bring data from
 	 */
 	private void getListing(final String uri) {
 		vectorResults = new Vector<Bookshare_Periodical_Bean>();
 
-		//Show progress bar
-		pd_spinning = ProgressDialog.show(this, null, resources.getString(R.string.fetching_periodicals), Boolean.TRUE);
+		// Show progress bar
+		pd_spinning = ProgressDialog.show(this, null,
+				resources.getString(R.string.fetching_periodicals),
+				Boolean.TRUE);
 
-		new Thread(){
-			public void run(){
-				try{
+		new Thread() {
+			public void run() {
+				try {
 					inputStream = bws.getResponseStream(password, uri);
 
-					// Once the response is obtained, send message to the handler
+					// Once the response is obtained, send message to the
+					// handler
 					Message msg = Message.obtain();
 					msg.what = DATA_FETCHED;
 					msg.setTarget(handler);
 					msg.sendToTarget();
-				}catch(IOException ioe){
+				} catch (IOException ioe) {
 					System.out.println(ioe);
-				}
-				catch(URISyntaxException use){
+				} catch (URISyntaxException use) {
 					System.out.println(use);
 				}
 			}
@@ -149,16 +148,16 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 
 	}
 
-
-	/**When user go to next page, this method clears the list and create the url
-	 *for the next page of results 
-	 * @param selectorId 
+	/**
+	 * When user go to next page, this method clears the list and create the url
+	 * for the next page of results
+	 * 
+	 * @param selectorId
 	 */
-	public void pageChangeSelected(int selectorId){
-		if(selectorId == NEXT_PAGE_BOOK_ID){
+	public void pageChangeSelected(int selectorId) {
+		if (selectorId == NEXT_PAGE_BOOK_ID) {
 			current_result_page++;
-		}
-		else if(selectorId == PREVIOUS_PAGE_BOOK_ID){
+		} else if (selectorId == PREVIOUS_PAGE_BOOK_ID) {
 			current_result_page--;
 		}
 		list.clear();
@@ -166,25 +165,26 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 		StringBuilder strBuilder = new StringBuilder(requestURI);
 		int index;
 
-		if((index = strBuilder.indexOf("?api_key=")) != -1){
+		if ((index = strBuilder.indexOf("?api_key=")) != -1) {
 			strBuilder.delete(index, strBuilder.length());
-			strBuilder.append("/page/"+current_result_page+"?api_key="+developerKey);
+			strBuilder.append("/page/" + current_result_page + "?api_key="
+					+ developerKey);
 		}
 		getListing(strBuilder.toString());
 	}
 
-
-	// Handler for dealing with the stream obtained as a result of search 
-	Handler handler = new Handler(){
+	// Handler for dealing with the stream obtained as a result of search
+	Handler handler = new Handler() {
 
 		@Override
-		public void handleMessage(Message msg){
+		public void handleMessage(Message msg) {
 
-			// Message received that data has been fetched from the bookshare web services 
-			if(msg.what == DATA_FETCHED){
+			// Message received that data has been fetched from the bookshare
+			// web services
+			if (msg.what == DATA_FETCHED) {
 
 				setContentView(R.layout.bookshare_menu_main);
-				searchET=(EditText)findViewById(R.id.searchText);								
+				searchET = (EditText) findViewById(R.id.searchText);
 				searchET.setSingleLine();
 				// Dismiss the progress dialog
 				pd_spinning.cancel();
@@ -192,21 +192,24 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 				String response_HTML = bws.convertStreamToString(inputStream);
 
 				// Cleanup the HTML formatted tags
-				String response = response_HTML.replace("&apos;", "\'").replace("&quot;", "\"").replace("&amp;", "and").replace("&#xd;","").replace("&#x97;", "-");
+				String response = response_HTML.replace("&apos;", "\'")
+						.replace("&quot;", "\"").replace("&amp;", "and")
+						.replace("&#xd;", "").replace("&#x97;", "-");
 
 				System.out.println(response);
 				// Parse the response of search result
 				parseResponse(response);
 
-				//process list
+				// process list
 				list.clear();
 
-				// For each bean object stored in the vector, create a row in the list
-				for(Bookshare_Periodical_Bean bean : vectorResults){
+				// For each bean object stored in the vector, create a row in
+				// the list
+				for (Bookshare_Periodical_Bean bean : vectorResults) {
 
 					TreeMap<String, Object> row_item = new TreeMap<String, Object>();
 
-					//puts item title
+					// puts item title
 					row_item.put("title", bean.getTitle());
 
 					row_item.put("icon", R.drawable.periodicals);
@@ -215,70 +218,79 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 					list.add(row_item);
 				}
 
-				if(current_result_page > 1 ){
-					createPageChanger("Previous Page", PREVIOUS_PAGE_BOOK_ID, R.drawable.arrow_left_blue);
+				if (current_result_page > 1) {
+					createPageChanger("Previous Page", PREVIOUS_PAGE_BOOK_ID,
+							R.drawable.arrow_left_blue);
 				}
 
-				if(current_result_page < total_pages_result ){
-					createPageChanger("Next Page", NEXT_PAGE_BOOK_ID, R.drawable.arrow_right_blue);
+				if (current_result_page < total_pages_result) {
+					createPageChanger("Next Page", NEXT_PAGE_BOOK_ID,
+							R.drawable.arrow_right_blue);
 				}
 			}
 
-
 			// Instantiate the custom SimpleAdapter for populating the ListView
-			// The bookId view in the layout file is used to store the id , but is not shown on screen
+			// The bookId view in the layout file is used to store the id , but
+			// is not shown on screen
 			final MySimpleAdapter simpleadapter = new MySimpleAdapter(
-					getApplicationContext(),list,
-					R.layout.bookshare_menu_item);
-			//Set the adapter for this view
+					getApplicationContext(), list, R.layout.bookshare_menu_item);
+			// Set the adapter for this view
 			setListAdapter(simpleadapter);
 
 			searchET.addTextChangedListener(new TextWatcher() {
 
-                @Override
-                public void onTextChanged(CharSequence s, int start, int before, int count) {
-                    // Call back the Adapter with current character to Filter
-                    simpleadapter.getFilter().filter(s.toString());
-                }
+				@Override
+				public void onTextChanged(CharSequence s, int start,
+						int before, int count) {
+					// Call back the Adapter with current character to Filter
+					simpleadapter.getFilter().filter(s.toString());
+				}
 
-                @Override
-                public void beforeTextChanged(CharSequence s, int start, int count,int after) {
-                }
+				@Override
+				public void beforeTextChanged(CharSequence s, int start,
+						int count, int after) {
+				}
 
-                @Override
-                public void afterTextChanged(Editable s) {
-                }
-            });
-			
+				@Override
+				public void afterTextChanged(Editable s) {
+				}
+			});
+
 			ListView lv = getListView();
 			lv.setTextFilterEnabled(true);
-			
+
 			View decorView = getWindow().getDecorView();
 			if (null != decorView) {
 				String resultsMessage;
 				if (vectorResults.isEmpty()) {
-					resultsMessage = resources.getString(R.string.search_complete_no_books);
-					setResult(InternalReturnCodes.NO_BOOKS_FOUND);
-					confirmAndClose("no books found", 3000);
+					resultsMessage = resources
+							.getString(R.string.search_complete_no_periodicals);
+					setResult(InternalReturnCodes.NO_PERIODICALS_FOUND);
+					confirmAndClose("no periodicals found", 3000);
 				} else {
-					resultsMessage = resources.getString(R.string.search_complete_with_books);
+					resultsMessage = resources
+							.getString(R.string.search_complete_with_periodicals);
 				}
 				decorView.setContentDescription(resultsMessage);
 			}
 
-			//When list item is clicked item should do a lookup and return editions/revisions
-			//they have on the particular magazine (thush)
-			lv.setOnItemClickListener(new OnItemClickListener(){
+			// When list item is clicked item should do a lookup and return
+			// editions/revisions
+			// they have on the particular magazine (thush)
+			lv.setOnItemClickListener(new OnItemClickListener() {
 
-				public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+				public void onItemClick(AdapterView<?> parent, View view,
+						int position, long id) {
 
 					// Obtain the layout for selected row
-					LinearLayout row_view  = (LinearLayout)view;
+					LinearLayout row_view = (LinearLayout) view;
 
-					//Obtain the book ID
-					TextView bookId = (TextView)row_view.findViewById(R.id.bookId);
-					if (null != bookId.getText().toString() ) {
-						int numericId =  Integer.valueOf(bookId.getText().toString());
+					// Obtain the book ID
+					TextView bookId = (TextView) row_view
+							.findViewById(R.id.bookId);
+					if (null != bookId.getText().toString()) {
+						int numericId = Integer.valueOf(bookId.getText()
+								.toString());
 						if (numericId < 0) {
 							pageChangeSelected(numericId);
 							return;
@@ -286,18 +298,25 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 					}
 
 					// Find the corresponding bean object for this row
-					for(Bookshare_Periodical_Bean bean : vectorResults){
+					for (Bookshare_Periodical_Bean bean : vectorResults) {
 
-						// Since book ID is unique, that can serve as comparison parameter
+						// Since book ID is unique, that can serve as comparison
+						// parameter
 						// Retrieve the book ID from the entry that is clicked
-						if(bean.getId().equalsIgnoreCase(bookId.getText().toString())){
+						if (bean.getId().equalsIgnoreCase(
+								bookId.getText().toString())) {
 							String bookshare_ID = bean.getId();
-							Intent intent = new Intent(getApplicationContext(),Bookshare_Periodical_Edition_Listing.class);
+							Intent intent = new Intent(getApplicationContext(),
+									Bookshare_Periodical_Edition_Listing.class);
 							String uri;
-							if(isFree)
-								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH + bookshare_ID + "?api_key="+developerKey;
+							if (isFree)
+								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH
+										+ bookshare_ID + "?api_key="
+										+ developerKey;
 							else
-								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH + bookshare_ID +"/for/"+username+"?api_key="+developerKey;
+								uri = URI_BOOKSHARE_PERIODICAL_EDITION_SEARCH
+										+ bookshare_ID + "/for/" + username
+										+ "?api_key=" + developerKey;
 
 							/*--------------------------------------------------------------------
 							if((isFree && bean.getAvailableToDownload().equals("1") &&
@@ -309,21 +328,21 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 								intent.putExtra("isDownloadable", false);
 							}
 							--------------------------------------------------------------------*/
-							
+
 							intent.putExtra("ID_SEARCH_URI", uri);
-							if(!isFree){
+							if (!isFree) {
 								intent.putExtra("username", username);
 								intent.putExtra("password", password);
 							}
 
-							startActivityForResult(intent, START_BOOKSHARE_PERIODICAL_EDITION_ACTIVITY);
+							startActivityForResult(intent,
+									START_BOOKSHARE_PERIODICAL_EDITION_ACTIVITY);
 							break;
 						}
 					}
 				}
 			});
 		}
-
 
 		private void createPageChanger(String title, int id, int iconId) {
 			TreeMap<String, Object> row_item = new TreeMap<String, Object>();
@@ -335,39 +354,39 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 			list.add(row_item);
 		}
 	};
-	
+
 	@Override
-	protected void onActivityResult(int requestCode, int resultCode, Intent data){
-		if(requestCode == START_BOOKSHARE_PERIODICAL_EDITION_ACTIVITY){
-			if(resultCode == BOOKSHARE_PERIODICAL_EDITION_FINISHED){
+	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+		if (requestCode == START_BOOKSHARE_PERIODICAL_EDITION_ACTIVITY) {
+			if (resultCode == BOOKSHARE_PERIODICAL_EDITION_FINISHED) {
 				setResult(BOOKSHARE_PERIODICAL_LISTING_FINISHED);
 				finish();
-			} else if (resultCode == InternalReturnCodes.NO_BOOK_FOUND) {
-                setResult(resultCode);
-                finish();
-            }
+			} else if (resultCode == InternalReturnCodes.NO_PERIODICAL_FOUND) {
+				setResult(resultCode);
+				finish();
+			}
 		}
 	}
 
-
-    /*
-     * Display voiceable message and then close
-     */
-    private void confirmAndClose(String msg, int timeout) {
-        final ParentCloserDialog dialog = new ParentCloserDialog(this, this);
-        dialog.popup(msg, timeout);
-    }
-    
-
-    /**
-	 * Uses a SAX parser to parse the response
-	 * @param response String representing the response
+	/*
+	 * Display voiceable message and then close
 	 */
-	private void parseResponse(String response){
+	private void confirmAndClose(String msg, int timeout) {
+		final ParentCloserDialog dialog = new ParentCloserDialog(this, this);
+		dialog.popup(msg, timeout);
+	}
+
+	/**
+	 * Uses a SAX parser to parse the response
+	 * 
+	 * @param response
+	 *            String representing the response
+	 */
+	private void parseResponse(String response) {
 
 		InputSource is = new InputSource(new StringReader(response));
 
-		try{
+		try {
 			/* Get a SAXParser from the SAXPArserFactory. */
 			SAXParserFactory spf = SAXParserFactory.newInstance();
 			SAXParser sp;
@@ -377,20 +396,17 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 			XMLReader parser = sp.getXMLReader();
 			parser.setContentHandler(new SAXHandler());
 			parser.parse(is);
-		}
-		catch(SAXException e){
+		} catch (SAXException e) {
 			System.out.println(e);
-		}
-		catch (ParserConfigurationException e) {
+		} catch (ParserConfigurationException e) {
 			System.out.println(e);
-		}
-		catch(IOException ioe){
+		} catch (IOException ioe) {
 			System.out.println(ioe);
 		}
 	}
 
 	// Class containing the logic for parsing the response of search results
-	private class SAXHandler extends DefaultHandler{
+	private class SAXHandler extends DefaultHandler {
 
 		boolean result = false;
 		boolean id = false;
@@ -399,110 +415,111 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 
 		Bookshare_Periodical_Bean result_bean;
 
-		public void startElement(String namespaceURI, String localName, String qName, Attributes atts){
-			
-			if(!total_pages_count_known)
-			{
-				if(qName.equalsIgnoreCase("num-pages")){
+		public void startElement(String namespaceURI, String localName,
+				String qName, Attributes atts) {
+
+			if (!total_pages_count_known) {
+				if (qName.equalsIgnoreCase("num-pages")) {
 					num_pages = true;
 					total_pages_count_known = false;
 				}
 			}
 
-			if(qName.equalsIgnoreCase("result")){
+			if (qName.equalsIgnoreCase("result")) {
 				result = true;
 				result_bean = new Bookshare_Periodical_Bean();
 			}
-			if(qName.equalsIgnoreCase("id")){
+			if (qName.equalsIgnoreCase("id")) {
 				id = true;
 			}
-			if(qName.equalsIgnoreCase("title")){
+			if (qName.equalsIgnoreCase("title")) {
 				title = true;
 			}
-			
+
 		}
 
-		public void endElement(String uri, String localName, String qName){
+		public void endElement(String uri, String localName, String qName) {
 
-			if(num_pages){
-				if(qName.equalsIgnoreCase("num-pages")){
+			if (num_pages) {
+				if (qName.equalsIgnoreCase("num-pages")) {
 					num_pages = false;
 				}
 			}
-			if(qName.equalsIgnoreCase("result")){
+			if (qName.equalsIgnoreCase("result")) {
 				result = false;
 				vectorResults.add(result_bean);
 				result_bean = null;
 			}
-			if(qName.equalsIgnoreCase("id")){
+			if (qName.equalsIgnoreCase("id")) {
 				id = false;
 			}
-			if(qName.equalsIgnoreCase("title")){
+			if (qName.equalsIgnoreCase("title")) {
 				title = false;
-			}			
+			}
 		}
 
-		public void characters(char[] c, int start, int length){
-			
-			if(num_pages){
-				total_pages_result = Integer.parseInt(new String(c,start,length));
+		public void characters(char[] c, int start, int length) {
+
+			if (num_pages) {
+				total_pages_result = Integer.parseInt(new String(c, start,
+						length));
 			}
-			if(result){
-				if(id){
-					result_bean.setId(new String(c,start,length));
+			if (result) {
+				if (id) {
+					result_bean.setId(new String(c, start, length));
 				}
-				if(title){
-					result_bean.setTitle(new String(c,start,length));
-				}				
+				if (title) {
+					result_bean.setTitle(new String(c, start, length));
+				}
 			}
 		}
 	}
 
-
 	// A custom ArrayAdapter class for providing data to the ListView
 	// Here we need to filter items by the text entered in the Edittext above
-	// Only the Periodical which contains the string entered in the edittext must be shown to the user. 
-	private class MySimpleAdapter extends ArrayAdapter implements Filterable{
-		
+	// Only the Periodical which contains the string entered in the edittext
+	// must be shown to the user.
+	private class MySimpleAdapter extends ArrayAdapter implements Filterable {
+
 		private List<TreeMap<String, Object>> allItemsArray;
-	    private List<TreeMap<String, Object>> filteredItemsArray;
-	    private Activity context;
-	    private TitleFilter filter;
-	    private LayoutInflater inflator;
-	    
-		public MySimpleAdapter(Context context, List<TreeMap<String, Object>> data,
-				int resource) {
+		private List<TreeMap<String, Object>> filteredItemsArray;
+		private Activity context;
+		private TitleFilter filter;
+		private LayoutInflater inflator;
+
+		public MySimpleAdapter(Context context,
+				List<TreeMap<String, Object>> data, int resource) {
 			super(context, resource, data);
-			allItemsArray=new ArrayList();
+			allItemsArray = new ArrayList();
 			allItemsArray.addAll(data);
-			filteredItemsArray=new ArrayList<TreeMap<String,Object>>();
+			filteredItemsArray = new ArrayList<TreeMap<String, Object>>();
 			filteredItemsArray.addAll(allItemsArray);
 		}
 
-		
 		@Override
-	    public Filter getFilter() {
-	        if (filter == null){
-	          filter=new TitleFilter();	        }
-	        
-	        
-	        return filter;
-	      }
-		
+		public Filter getFilter() {
+			if (filter == null) {
+				filter = new TitleFilter();
+			}
+
+			return filter;
+		}
+
 		/*
-		 * Retrieves view for the item in the adapter, at the
-		 * specified position and populates it with data.
+		 * Retrieves view for the item in the adapter, at the specified position
+		 * and populates it with data.
 		 */
 		@Override
 		public View getView(int position, View convertView, ViewGroup parent) {
 			if (convertView == null) {
-				convertView = getLayoutInflater().inflate(R.layout.bookshare_menu_item, null);
+				convertView = getLayoutInflater().inflate(
+						R.layout.bookshare_menu_item, null);
 			}
 
 			TreeMap<String, Object> data = (TreeMap<String, Object>) getItem(position);
 
 			((TextView) convertView.findViewById(R.id.text1))
-			.setText((String) data.get("title"));
+					.setText((String) data.get("title"));
 
 			/*----------------------------------------------------------------------------
 			StringBuilder authorsBuilder = new StringBuilder("");
@@ -521,67 +538,64 @@ public class Bookshare_Periodical_Listing extends ListActivity{
 			((ImageView) convertView.findViewById(R.id.row_icon))
 			.setImageResource((Integer) data.get("icon"));
 			------------------------------------------------------------------------------*/
-			
-			
-			if(data.get("download_icon") != null){
-				((ImageView)convertView.findViewById(R.id.bookshare_download_icon))
-				.setImageResource((Integer) data.get("download_icon"));
+
+			if (data.get("download_icon") != null) {
+				((ImageView) convertView
+						.findViewById(R.id.bookshare_download_icon))
+						.setImageResource((Integer) data.get("download_icon"));
 
 				((TextView) convertView.findViewById(R.id.bookId))
-				.setText((String) data.get("book_id"));
+						.setText((String) data.get("book_id"));
 			}
-			
+
 			((TextView) convertView.findViewById(R.id.bookId))
-			.setText((String) data.get("book_id"));
-			
+					.setText((String) data.get("book_id"));
+
 			return convertView;
 		}
-		
-		 private class TitleFilter extends Filter
-	        {
 
-			 	//do the filtering process
-	            @Override
-	            protected FilterResults performFiltering(CharSequence constraint) {
-	                
-	                constraint = constraint.toString().toLowerCase();
-	                FilterResults result = new FilterResults();
-	                if(constraint != null && constraint.toString().length() > 0)
-	                {
-	                    ArrayList<TreeMap<String, Object>> filteredItems = new ArrayList<TreeMap<String, Object>>();
-	                   
-	                    for(int i = 0, l = allItemsArray.size(); i < l; i++)
-	                    {
-	                    	TreeMap<String, Object> periodical = allItemsArray.get(i);
-	                        if(periodical.get("title").toString().toLowerCase().contains(constraint))
-	                            filteredItems.add(periodical);
-	                    }
-	                    result.count = filteredItems.size();
-	                    result.values = filteredItems;
-	                }
-	                else
-	                {
-	                    synchronized(this)
-	                    {
-	                        result.values = allItemsArray;
-	                        result.count = allItemsArray.size();
-	                    }
-	                }
-	                return result;
-	            }
+		private class TitleFilter extends Filter {
 
-	            //add results to filteredItemList
-	            @SuppressWarnings("unchecked")
-	            @Override
-	            protected void publishResults(CharSequence constraint, FilterResults results) {
-	               
-	            	filteredItemsArray = (ArrayList<TreeMap<String, Object>>)results.values;
-	                notifyDataSetChanged();
-	                clear();
-	                for(int i = 0, l = filteredItemsArray.size(); i < l; i++)
-	                    add(filteredItemsArray.get(i));
-	                notifyDataSetInvalidated();
-	            }
-	       }
+			// do the filtering process
+			@Override
+			protected FilterResults performFiltering(CharSequence constraint) {
+
+				constraint = constraint.toString().toLowerCase();
+				FilterResults result = new FilterResults();
+				if (constraint != null && constraint.toString().length() > 0) {
+					ArrayList<TreeMap<String, Object>> filteredItems = new ArrayList<TreeMap<String, Object>>();
+
+					for (int i = 0, l = allItemsArray.size(); i < l; i++) {
+						TreeMap<String, Object> periodical = allItemsArray
+								.get(i);
+						if (periodical.get("title").toString().toLowerCase()
+								.contains(constraint))
+							filteredItems.add(periodical);
+					}
+					result.count = filteredItems.size();
+					result.values = filteredItems;
+				} else {
+					synchronized (this) {
+						result.values = allItemsArray;
+						result.count = allItemsArray.size();
+					}
+				}
+				return result;
+			}
+
+			// add results to filteredItemList
+			@SuppressWarnings("unchecked")
+			@Override
+			protected void publishResults(CharSequence constraint,
+					FilterResults results) {
+
+				filteredItemsArray = (ArrayList<TreeMap<String, Object>>) results.values;
+				notifyDataSetChanged();
+				clear();
+				for (int i = 0, l = filteredItemsArray.size(); i < l; i++)
+					add(filteredItemsArray.get(i));
+				notifyDataSetInvalidated();
+			}
+		}
 	}
 }
