+++ src/org/geometerplus/android/fbreader/network/bookshare/InternalReturnCodes.java
@@ -1,13 +1,12 @@
 package org.geometerplus.android.fbreader.network.bookshare;
 
 /**
- * @author roms
- *         Date: 2/24/12
+ * @author roms Date: 2/24/12
  */
-public class InternalReturnCodes { 
-    
-    public static final int NO_BOOKS_FOUND = 10;
-    public static final int NO_BOOK_FOUND = 11;
-    public static final int NO_PERIODICAL_EDITION_FOUND = 12;
-    public static final int NO_PERIODICAL_EDITIONS_FOUND = 13;
+public class InternalReturnCodes {
+
+	public static final int NO_BOOKS_FOUND = 10;
+	public static final int NO_BOOK_FOUND = 11;
+	public static final int NO_PERIODICAL_FOUND = 12;
+	public static final int NO_PERIODICALS_FOUND = 13;
 }
