+++ src/org/geometerplus/android/fbreader/network/bookshare/subscription/BooksharePeriodicalDataSource.java
@@ -1,4 +1,4 @@
-package org.geometerplus.android.fbreader.subscription;
+package org.geometerplus.android.fbreader.network.bookshare.subscription;
 
 import java.util.ArrayList;
 import java.util.List;
