+++ src/org/geometerplus/android/fbreader/network/bookshare/subscription/Bookshare_Subscription_Download_Service.java
@@ -1,4 +1,4 @@
-package org.geometerplus.android.fbreader.network.bookshare;
+package org.geometerplus.android.fbreader.network.bookshare.subscription;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
@@ -6,10 +6,12 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.Vector;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -24,41 +26,37 @@ import org.benetech.android.R;
 import org.bookshare.net.BookshareWebservice;
 import org.geometerplus.android.fbreader.FBReader;
 import org.geometerplus.android.fbreader.network.BookDownloaderService;
-import org.geometerplus.android.fbreader.subscription.AllDbPeriodicalEntity;
-import org.geometerplus.android.fbreader.subscription.BooksharePeriodicalDataSource;
-import org.geometerplus.android.fbreader.subscription.PeriodicalDBUtils;
-import org.geometerplus.android.fbreader.subscription.PeriodicalsSQLiteHelper;
-import org.geometerplus.android.fbreader.subscription.SubscribedDbPeriodicalEntity;
+import org.geometerplus.android.fbreader.network.bookshare.BookshareDeveloperKey;
+import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Edition_Metadata_Bean;
+import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Error_Bean;
+import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Periodical_Edition_Bean;
+import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Webservice_Login;
 import org.geometerplus.fbreader.Paths;
+import org.geometerplus.fbreader.fbreader.FBReaderApp.AutomaticDownloadType;
 import org.geometerplus.zlibrary.core.filesystem.ZLFile;
 import org.geometerplus.zlibrary.core.resources.ZLResource;
 
-import android.app.IntentService;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
+import android.app.Service;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
 import android.os.AsyncTask;
+import android.os.Binder;
 import android.os.Handler;
+import android.os.IBinder;
 import android.os.Message;
 import android.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.Log;
 import android.widget.RemoteViews;
 
-/**
- * This is an intent service. This was used to stop simultaneous downloas as it
- * causes system not to respond as well as corrupt zip files. So with this
- * downloads are done sequentially
- * 
- * @author thushan
- * 
- */
-public class SubscriptionDownloadService extends IntentService {
+public class Bookshare_Subscription_Download_Service extends Service {
 
+	// SubscriptionSQLiteHelper dbHelper;
 	BooksharePeriodicalDataSource dataSource;
 	PeriodicalsSQLiteHelper dbHelper;
 	private String username;
@@ -67,25 +65,73 @@ public class SubscriptionDownloadService extends IntentService {
 	private String omDownloadPassword;
 	private boolean isFree = false;
 	private boolean isOM;
-	private boolean downloadSuccess;
-
 	private String developerKey = BookshareDeveloperKey.DEVELOPER_KEY;
 
 	private Bookshare_Edition_Metadata_Bean metadata_bean;
 	private Set<Integer> myOngoingNotifications = new HashSet<Integer>();
 	private SQLiteDatabase periodicalDb;
+	private PeriodicalEditionListFetcher editionFetcher;
+	private PeriodicalEditionMetadataFetcher metadataFetcher;
+	private ServiceBinder serviceBinder = new ServiceBinder();
 
-	private Bookshare_Error_Bean error;
-	final BookshareWebservice bws = new BookshareWebservice(
-			Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
+	@Override
+	public IBinder onBind(Intent arg0) {
 
-	public SubscriptionDownloadService() {
-		super("SubscriptionDownloadService");
+		SharedPreferences logingPrefs = PreferenceManager
+				.getDefaultSharedPreferences(getApplicationContext());
+		username = logingPrefs.getString("username", "");
+		password = logingPrefs.getString("password", "");
+
+		// TODO: Get first name, last name, member id
+		dataSource = BooksharePeriodicalDataSource
+				.getInstance(getApplicationContext());
+		dbHelper = new PeriodicalsSQLiteHelper(getApplicationContext());
+		periodicalDb = dbHelper.getWritableDatabase();
+
+		editionFetcher = new PeriodicalEditionListFetcher();
+
+		if (username == null || password == null || TextUtils.isEmpty(username)) {
+			isFree = true;
+		}
+
+		return serviceBinder;
+	}
+
+	@Override
+	public boolean onUnbind(Intent intent) {
+		stopSelf();
+		periodicalDb.close();
+		return super.onUnbind(intent);
+	}
+
+	@Override
+	public void onCreate() {
+		super.onCreate();
+
+		// dbHelper=new SubscriptionSQLiteHelper(getApplicationContext());
+		// dataSrc=new BooksharePeriodicalDataSource(getApplicationContext());
 
 	}
 
 	@Override
-	protected void onHandleIntent(Intent intent) {
+	public void onDestroy() {
+		super.onDestroy();
+		periodicalDb.close();
+	}
+
+	@Override
+	public int onStartCommand(Intent intent, int flags, int startId) {
+		Log.i(getClass().getName(),
+				"**** Service Started by Alarm Manager ****");
+
+		if (serviceBinder == null) {
+			serviceBinder = new ServiceBinder();
+		}
+
+		SharedPreferences logingPrefs = PreferenceManager
+				.getDefaultSharedPreferences(getApplicationContext());
+		username = logingPrefs.getString("username", "");
+		password = logingPrefs.getString("password", "");
 
 		dataSource = BooksharePeriodicalDataSource
 				.getInstance(getApplicationContext());
@@ -100,297 +146,200 @@ public class SubscriptionDownloadService extends IntentService {
 			}
 		}
 
-		username = intent.getStringExtra("username");
-		password = intent.getStringExtra("password");
-
 		if (username == null || password == null || TextUtils.isEmpty(username)) {
 			isFree = true;
 		}
 
-		metadata_bean = (Bookshare_Edition_Metadata_Bean) intent
-				.getSerializableExtra("metadata_bean");
+		ArrayList<String> ids = (ArrayList<String>) intent
+				.getStringArrayListExtra(FBReader.SUBSCRIBED_PERIODICAL_IDS_KEY);
 
-		Log.i(getClass().getName(),
-				"Downloding one by one... " + "Title: "
-						+ metadata_bean.getTitle() + " Edition: "
-						+ metadata_bean.getEdition());
-		// new DownloadFilesTask().execute();
-
-		// From here onwards till method's end it's similar to
-		// DownloadFilesTask() Task
-		final String id = metadata_bean.getContentId();
-		String download_uri;
-		if (isFree)
-			download_uri = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
-					+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST
-					+ "/download/content/" + id + "/version/1?api_key="
-					+ developerKey;
-		// TODO: Uncomment & Implement
-		/*
-		 * else if(isOM){
-		 * 
-		 * download_uri = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL +
-		 * Bookshare_Webservice_Login.BOOKSHARE_API_HOST + "/download/member/"
-		 * +memberId+"content/"+id+"/version/1/for/"+username
-		 * +"?api_key="+developerKey; }
-		 */
-
-		else {
-			download_uri = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
-					+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST
-					+ "/download/content/" + id + "/version/1/for/" + username
-					+ "?api_key=" + developerKey;
+		String downTypeStr = intent
+				.getStringExtra(FBReader.AUTOMATIC_DOWNLOAD_TYPE_KEY);
+		AutomaticDownloadType downType;
+		Log.i(getClass().getName(), "Extras passed: id array size" + ids.size()
+				+ " , " + downTypeStr);
+		// Determine the download type user has set
+		if (AutomaticDownloadType.downloadAll.toString().equals(downTypeStr)) {
+			downType = AutomaticDownloadType.downloadAll;
+		} else {
+			downType = AutomaticDownloadType.downloadMostRecent;
 		}
 
-		final Notification progressNotification = createDownloadProgressNotification(metadata_bean
-				.getTitle());
+		if (ids != null && ids.size() > 0) {
+			for (String id : ids) {
+				Log.i(getClass().getName(),
+						"Periodical search started by alarm: " + id);
+				serviceBinder.getUpdates(downType, id);
 
-		final NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
-		myOngoingNotifications.add(Integer.valueOf(id));
-		notificationManager.notify(Integer.valueOf(id), progressNotification);
-
-		try {
-			System.out.println("download_uri :" + download_uri);
-			HttpResponse response = bws.getHttpResponse(password, download_uri);
-			// Get hold of the response entity
-			HttpEntity entity = response.getEntity();
-
-			if (entity != null) {
-				String filename = "bookshare_" + Math.random() * 10000 + ".zip";
-				if (metadata_bean.getTitle() != null
-						&& metadata_bean.getEdition() != null) {
-					String temp = "";
-					// Changed the file name to <title>_<edition>
-					temp = metadata_bean.getTitle() + "_"
-							+ metadata_bean.getEdition();
-
-					filename = temp;
-					filename = filename.replaceAll(" +", "_").replaceAll(":",
-							"__");
-					if (isOM) {
-						// TODO: Uncomment & Implement
-						// filename = filename + "_" + firstName + "_" +
-						// lastName;
-					}
-				}
-				String zip_file = Paths.BooksDirectoryOption().getValue() + "/"
-						+ filename + ".zip";
-				downloadedBookDir = Paths.BooksDirectoryOption().getValue()
-						+ "/" + filename;
-
-				File downloaded_zip_file = new File(zip_file);
-				if (downloaded_zip_file.exists()) {
-					downloaded_zip_file.delete();
-				}
-				Header header = entity.getContentType();
-				// Log.w("FBR", "******  zip_file *****" + zip_file);
-				final String headerValue = header.getValue();
-				if (headerValue.contains("zip") || headerValue.contains("bks2")) {
-					try {
-						System.out.println("Contains zip");
-						java.io.BufferedInputStream in = new java.io.BufferedInputStream(
-								entity.getContent());
-						java.io.FileOutputStream fos = new java.io.FileOutputStream(
-								downloaded_zip_file);
-						java.io.BufferedOutputStream bout = new BufferedOutputStream(
-								fos, 1024);
-						byte[] data = new byte[1024];
-						int x = 0;
-						while ((x = in.read(data, 0, 1024)) >= 0) {
-							bout.write(data, 0, x);
-						}
-						fos.flush();
-						bout.flush();
-						fos.close();
-						bout.close();
-						in.close();
-
-						System.out.println("******** Downloading complete");
-
-						// Unzip the encrypted archive file
-						if (!isFree) {
-							System.out
-									.println("******Before creating ZipFile******"
-											+ zip_file);
-							// Initiate ZipFile object with the path/name of
-							// the zip file.
-							ZipFile zipFile = new ZipFile(zip_file);
-
-							// Check to see if the zip file is password
-							// protected
-							if (zipFile.isEncrypted()) {
-								System.out.println("******isEncrypted******");
-
-								// if yes, then set the password for the zip
-								// file
-								if (!isOM) {
-									zipFile.setPassword(password);
-								}
-								// Set the OM password sent by the Intent
-								else {
-									// Obtain the SharedPreferences object
-									// shared across the application. It is
-									// stored in login activity
-									SharedPreferences login_preference = PreferenceManager
-											.getDefaultSharedPreferences(getApplicationContext());
-									omDownloadPassword = login_preference
-											.getString("downloadPassword", "");
-									zipFile.setPassword(omDownloadPassword);
-								}
-							}
-
-							// Get the list of file headers from the zip
-							// file
-							List fileHeaderList = zipFile.getFileHeaders();
-
-							System.out.println("******Before for******");
-							// Loop through the file headers
-							for (int i = 0; i < fileHeaderList.size(); i++) {
-								FileHeader fileHeader = (FileHeader) fileHeaderList
-										.get(i);
-								System.out.println(downloadedBookDir);
-								// Extract the file to the specified
-								// destination
-								zipFile.extractFile(fileHeader,
-										downloadedBookDir);
-							}
-						}
-						// Unzip the non-encrypted archive file
-						else {
-							try {
-								File file = new File(downloadedBookDir);
-								file.mkdir();
-								String destinationname = downloadedBookDir
-										+ "/";
-								byte[] buf = new byte[1024];
-								ZipInputStream zipinputstream = null;
-								ZipEntry zipentry;
-								zipinputstream = new ZipInputStream(
-										new FileInputStream(zip_file));
-
-								zipentry = zipinputstream.getNextEntry();
-								while (zipentry != null) {
-									// for each entry to be extracted
-									String entryName = zipentry.getName();
-									System.out
-											.println("entryname " + entryName);
-									int n;
-									FileOutputStream fileoutputstream;
-									File newFile = new File(entryName);
-									String directory = newFile.getParent();
-
-									if (directory == null) {
-										if (newFile.isDirectory())
-											break;
-									}
-
-									fileoutputstream = new FileOutputStream(
-											destinationname + entryName);
-
-									while ((n = zipinputstream.read(buf, 0,
-											1024)) > -1)
-										fileoutputstream.write(buf, 0, n);
-
-									fileoutputstream.close();
-									zipinputstream.closeEntry();
-									zipentry = zipinputstream.getNextEntry();
-
-								}// while
-
-								zipinputstream.close();
-							} catch (Exception e) {
-								e.printStackTrace();
-							}
-						}
-						// Delete the downloaded zip file as it has been
-						// extracted
-						downloaded_zip_file = new File(zip_file);
-						if (downloaded_zip_file.exists()) {
-							downloaded_zip_file.delete();
-						}
-						downloadSuccess = true;
-					} catch (ZipException e) {
-						Log.e("FBR", "Zip Exception", e);
-					}
-				} else {
-					downloadSuccess = false;
-					error = new Bookshare_Error_Bean();
-					error.parseInputStream(response.getEntity().getContent());
-				}
-			}
-		} catch (URISyntaxException use) {
-			System.out.println("URISyntaxException: " + use);
-		} catch (IOException ie) {
-			System.out.println("IOException: " + ie);
-		}
-
-		if (downloadSuccess) {
-			// Get download time/date
-			Calendar currCal = Calendar.getInstance();
-			String currentDate = currCal.get(Calendar.MONTH) + "/"
-					+ currCal.get(Calendar.DATE) + "/"
-					+ currCal.get(Calendar.YEAR) + "";
-			String currentTime = currCal.get(Calendar.HOUR_OF_DAY) + ":"
-					+ currCal.get(Calendar.MINUTE) + ":"
-					+ currCal.get(Calendar.SECOND) + "";
-
-			// create alldb and subscribeddb entities to be inserted to
-			// their respective dbs
-			AllDbPeriodicalEntity allEntity = new AllDbPeriodicalEntity(
-					metadata_bean.getPeriodicalId(), metadata_bean.getTitle(),
-					metadata_bean.getEdition(), Integer.parseInt(metadata_bean
-							.getRevision()), currentDate, currentTime);
-			SubscribedDbPeriodicalEntity subEntity = new SubscribedDbPeriodicalEntity(
-					metadata_bean.getPeriodicalId(), metadata_bean.getTitle(),
-					metadata_bean.getEdition(), Integer.parseInt(metadata_bean
-							.getRevision()));
-			dataSource.insertEntity(periodicalDb,
-					PeriodicalsSQLiteHelper.TABLE_ALL_PERIODICALS, allEntity);
-			if (dataSource.doesExist(periodicalDb,
-					PeriodicalsSQLiteHelper.TABLE_SUBSCRIBED_PERIODICALS,
-					subEntity)) {
-				SubscribedDbPeriodicalEntity dbSubEntity = (SubscribedDbPeriodicalEntity) dataSource
-						.getEntity(
-								periodicalDb,
-								PeriodicalsSQLiteHelper.TABLE_SUBSCRIBED_PERIODICALS,
-								subEntity);
-				// Check whether we're actually trying to put a greater version
-				// than existing version
-				if (subEntity.getLatestEdition().equals(
-						PeriodicalDBUtils.getRecentEditionString(
-								subEntity.getLatestEdition(),
-								dbSubEntity.getLatestEdition()))) {
-					dataSource
-							.insertEntity(
-									periodicalDb,
-									PeriodicalsSQLiteHelper.TABLE_SUBSCRIBED_PERIODICALS,
-									subEntity);
+				try {
+					Thread.sleep(2000);
+				} catch (InterruptedException e) {
+					Log.e(getClass().getName(),
+							e.getCause() + " ," + e.getMessage());
 				}
 			}
 		} else {
+			Log.e(getClass().getName(),
+					"Couldn't find any subscribed Periodicals");
+		}
+		return super.onStartCommand(intent, flags, startId);
 
-			downloadedBookDir = null;
+	}
+
+	public class ServiceBinder extends Binder implements
+			IPeriodicalDownloadAPI, PeriodicalEditionListener,
+			PeriodicalMetadataListener {
+		AutomaticDownloadType downType;
+
+		@Override
+		public boolean downloadPeriodical(Bookshare_Edition_Metadata_Bean bean) {
+			metadata_bean = bean;
+
+			Intent downloadService = new Intent(
+					Bookshare_Subscription_Download_Service.this,
+					SubscriptionDownloadService.class);
+			downloadService.putExtra("username", username);
+			downloadService.putExtra("password", password);
+			downloadService.putExtra("metadata_bean", metadata_bean);
+			startService(downloadService);
+			// new DownloadFilesTask().execute();
+			return false;
 		}
 
-		final Handler downloadFinishHandler = new Handler() {
-			public void handleMessage(Message message) {
-				final NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
-				int id = Integer.valueOf(metadata_bean.getContentId());
-				notificationManager.cancel(id);
-				myOngoingNotifications.remove(Integer.valueOf(id));
-				File file = null;
-				if (downloadSuccess) {
-					file = new File(getOpfFile().getPath());
-				}
-				notificationManager.notify(
-						id,
-						createDownloadFinishNotification(file,
-								metadata_bean.getTitle(), message.what != 0));
-			}
-		};
+		@Override
+		public void getUpdates(AutomaticDownloadType downType, String id) {
 
-		downloadFinishHandler.sendEmptyMessage(downloadSuccess ? 1 : 0);
-		periodicalDb.close();
+			this.downType = downType;
+			String serviceURI = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
+					+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST
+					+ "/periodical/id/"
+					+ id
+					+ "/for/"
+					+ username
+					+ "?api_key="
+					+ developerKey;
+			Log.i(getClass().getName(),
+					"Fetching Periodical List for periodical with id: " + id
+							+ " for: " + username);
+
+			editionFetcher = new PeriodicalEditionListFetcher();
+			editionFetcher.getListing(serviceURI, password, this);
+
+		}
+
+		@Override
+		public void onPeriodicalEditionListResponse(
+				Vector<Bookshare_Periodical_Edition_Bean> results) {
+			if (results == null) {
+				Log.e(getClass().getName(),
+						"Couldn't fetch any periodical Editions");
+			} else {
+				Log.i(getClass().getName(),
+						"Found and Fetched " + results.size() + " periodicals");
+				ArrayList<AllDbPeriodicalEntity> entities = new ArrayList<AllDbPeriodicalEntity>();
+				for (Bookshare_Periodical_Edition_Bean bean : results) {
+					Log.i(getClass().getName(),
+							"Found and fetched periodical: Title "
+									+ bean.getTitle() + " Edition: "
+									+ bean.getEdition());
+					if (bean.getId() != null && bean.getEdition() != null
+							&& TextUtils.isDigitsOnly(bean.getRevision())) {
+						entities.add(new AllDbPeriodicalEntity(bean.getId(),
+								bean.getTitle(), bean.getEdition(), Integer
+										.parseInt(bean.getRevision()), null,
+								null));
+					}
+				}
+
+				// URL to request metadata of particular edition of a periodical
+				String serviceURI;
+				// if user has set settings to download only the most recent
+				if (downType == AutomaticDownloadType.downloadMostRecent) {
+					AllDbPeriodicalEntity maxEntity = PeriodicalDBUtils
+							.getMostRecentEdition(entities);
+
+					// download the periodical only if it's not been downloaded
+					// before
+					if (!dataSource.doesExist(periodicalDb,
+							PeriodicalsSQLiteHelper.TABLE_ALL_PERIODICALS,
+							maxEntity)) {
+						serviceURI = getEditionRequestURL(maxEntity);
+
+						metadataFetcher = new PeriodicalEditionMetadataFetcher(
+								maxEntity.getId(), maxEntity.getTitle());
+						metadataFetcher.getListing(serviceURI, password, this);
+					}
+				}
+				// If user has set settings to download all periodicals
+				else if (downType == AutomaticDownloadType.downloadAll) {
+					for (AllDbPeriodicalEntity entity : entities) {
+						// download the periodical only if it's not been
+						// downloaded before
+						// TODO: Download periodicals which is higer than the
+						// highest in alldbperiodicals
+						if (!dataSource.doesExist(periodicalDb,
+								PeriodicalsSQLiteHelper.TABLE_ALL_PERIODICALS,
+								entity)) {
+							serviceURI = getEditionRequestURL(entity);
+
+							metadataFetcher = new PeriodicalEditionMetadataFetcher(
+									entity.getId(), entity.getTitle());
+							metadataFetcher.getListing(serviceURI, password,
+									this);
+
+							// This is to reduce number of queries per second
+							try {
+								Thread.sleep(1000);
+							} catch (InterruptedException ex) {
+								Log.e(getClass().getName(), ex.getCause()
+										+ " ," + ex.getMessage());
+							}
+						}
+					}
+				}
+			}
+
+		}
+
+		private String getEditionRequestURL(AllDbPeriodicalEntity entity) {
+			String serviceURI = Bookshare_Webservice_Login.BOOKSHARE_API_PROTOCOL
+					+ Bookshare_Webservice_Login.BOOKSHARE_API_HOST
+					+ "/periodical/id/"
+					+ entity.getId()
+					+ "/edition/"
+					+ entity.getEdition()
+					+ "/revision/"
+					+ entity.getRevision()
+					+ "/for/" + username + "?api_key=" + developerKey;
+			return serviceURI;
+		}
+
+		@Override
+		public Bookshare_Edition_Metadata_Bean getDetails(
+				Bookshare_Periodical_Edition_Bean bean) {
+			if (bean == null) {
+				Log.e(getClass().getName(), "Couldn't obtain edition details");
+			} else {
+				Log.i(getClass().getName(),
+						"Fetched Periodical: " + bean.getId() + " "
+								+ bean.getTitle() + " " + bean.getEdition());
+			}
+			return null;
+		}
+
+		@Override
+		public void onPeriodicalMetadataResponse(
+				Bookshare_Edition_Metadata_Bean result) {
+			if (result == null) {
+				Log.e(getClass().getName(), "Couldn't obtain edition details");
+			} else {
+				downloadPeriodical(result);
+				Log.i(getClass().getName(),
+						"Fetched Periodical: " + result.getPeriodicalId() + " "
+								+ result.getTitle() + " " + result.getEdition());
+			}
+
+		}
 
 	}
 
@@ -399,7 +348,6 @@ public class SubscriptionDownloadService extends IntentService {
 		private Bookshare_Error_Bean error;
 		final BookshareWebservice bws = new BookshareWebservice(
 				Bookshare_Webservice_Login.BOOKSHARE_API_HOST);
-
 		private boolean downloadSuccess;
 
 		// Will be called in the UI thread
@@ -693,7 +641,6 @@ public class SubscriptionDownloadService extends IntentService {
 			};
 
 			downloadFinishHandler.sendEmptyMessage(downloadSuccess ? 1 : 0);
-			periodicalDb.close();
 		}
 	}
