+++ src/org/geometerplus/android/fbreader/network/bookshare/subscription/PeriodicalDBUtils.java
@@ -1,4 +1,4 @@
-package org.geometerplus.android.fbreader.subscription;
+package org.geometerplus.android.fbreader.network.bookshare.subscription;
 
 import java.util.Calendar;
 import java.util.List;
